FIX: handle crashes when all items on a display list are hidden
When we filter a list of docs to get only the latest doc, but the latest doc is set to be invisible, we get an error. I fixed it so that it takes the next in line which is visible. If all docs in a given product are invisible, it does not display anything for that product.