DOCS-2089 - Use VcsRootInstance instead of VcsRoot in the locator
When we used VcsRoot in the locator, some builds weren't found. Using VcsRootInstace solved the issue.
The official TeamCity documentation explains why we experienced this issue:

"VCS root is a setting configured in the TeamCity UI. VCS root instance is an internal TeamCity entity which is derived from the VCS root to perform the actual VCS operation. If a VCS root has no % -references to parameters, a single VCS root corresponds to a single VCS root instance. A single VCS root can generate several VCS root instances: for example, if a VCS root has % -reference to a parameter and the reference resolves to a different value, when the VCS root is attached to different configurations, or when a custom build is run."