DOCS-2111 - Fix redirecting from external sites
The authentication controller saves info about the requested page in req.session.redirectTo when the request is made. This way, the user can get to the target page
after they log in.
We do not want to redirect the user to an external page from which they opened the doc site link. Therefore, the gw-login route does not need to catch the referer and assign it
to req.session.redirectTo.