DOCS-2891 - Run steps that set AWS credentials in a Docker container
We now have a dedicated pool of agents. Therefore, when a build step doesn’t run in a Docker container and it sets AWS credentials, it overrides the S3 storage keys used for publishing artifacts.
I also removed all explicit pulls of the Docker image to reduce the data transfer.