DOCS-4937 - Fix root breadcrumbs
Fix the build for building and deploying the Docker image for React landing pages. The step for generating the root-breadcrumbs.json file was in the wrong place and the file wasn't copied to the Docker image.
Moved the step to a separate build and added it as an artifact dependency to the build that creates the Docker image for React landing pages.