DOCS-5196 - Don't use GUIDEWIRE-CORP IDP for GW employee login
Don't use the IDP parameter for GW employee login as it isn't needed anymore. GW Identity Hub login screens changed some time ago and now as an employee you don't have to provide your email twice when you log in through the GW Cloud option. Removal of the IDP parameter improves the quality and stability of the code - we don't have modify strategies on the fly to add the IDP parameter.