fix: increase probing timeouts
Readiness and liveness probes often fail on prod.
The server handles a lot of traffic
so the timeout of 1 second may be too short.
If the timeout is too short and the probe fails too many times,
the server will be restarted.
Increased the timeout to 5 seconds
to allow for delays in server response due to heavy traffic on prod.

Task-id: DOCT-568