fix: set NODE_ENV to production on all envs
We have a few conditions depending on the NODE_ENV variable.
Before,
the value of development was only set
when the server was running locally
(the value is hardcoded in the package.json script).
In Kubernetes, we didn't have it set at all.
We will use production for all envs.
development will be only used for local development.

Task-id: DOCT-590